# PERF: concat_same_type for PeriodDtype