xref #23362

```
import numpy as np
import pandas as pd

a = np.random.randint(2000, 2100, size=1000)
b = np.random.randint(2000, 2100, size=1000)

x = pd.core.arrays.period_array(a, freq='B')
y = pd.core.arrays.period_array(b, freq='B')

%timeit x._concat_same_type([x, y])
6.86 µs ± 75.2 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)  # <- main
4.24 µs ± 109 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)  # <- PR
```